#304 refactor(db): Update Routine alarm ID type to Int
This commit changes the data type of the `idAlarm` field in the `RoutineEntity` and the return type of `getIdAlarms` in `RoutineDao` from `Long` to `Int`. This ensures consistency in how alarm identifiers are handled within the routine database module.

**Key Changes:**

*   **`RoutineDao.kt`**:
    *   Updated `getIdAlarms()` to return a `List<Int>` instead of `List<Long>`.
*   **`RoutineEntity.kt`**:
    *   Changed the `idAlarm` property type from `Long?` to `Int?`.

+++ core/db/src/main/java/com/rahim/yadino/db/routine/dao/RoutineDao.kt
@@ -46,7 +46,7 @@ interface RoutineDao {
   suspend fun removeSampleRoutine()
 
   @Query("SELECT idAlarm FROM tbl_routine WHERE isChecked=0")
-  suspend fun getIdAlarms(): List<Long>
+  suspend fun getIdAlarms(): List<Int>
 
   @Query("SELECT EXISTS(SELECT 1 FROM tbl_routine WHERE isChecked = 0)")
   fun haveAlarm(): Flow<Boolean>
